+++ next.config.ts
@@ -1,4 +1,5 @@
 import type { NextConfig } from "next";
+import packageJson from "./package.json";
 
 const { PROD_URL, PORT, JWT_SECRET, JWT_DURATION, PASSWORD } = process.env;
 
@@ -7,7 +8,13 @@ const API_URL =
 
 const nextConfig: NextConfig = {
   reactStrictMode: true,
-  env: { API_URL, JWT_SECRET, JWT_DURATION, PASSWORD },
+  env: {
+    API_URL,
+    JWT_SECRET,
+    JWT_DURATION,
+    PASSWORD,
+    APP_VERSION: packageJson.version,
+  },
 };
 
 export default nextConfig;