+++ src/@types/category.ts
@@ -2,4 +2,5 @@ export type CategoryItem = {
   id: number;
   name: string;
   color: string;
+  position: number;
 };