+++ src/app/api/category/route.ts
@@ -1,0 +1,84 @@
+import { NextRequest, NextResponse } from "next/server";
+import { sql } from "@/services/database";
+
+import * as Auth from "@/services/auth";
+
+export async function POST(request: NextRequest) {
+  try {
+    await Auth.check(request);
+
+    const { position } = await request.json();
+
+    await sql.execute(`INSERT INTO category (position) VALUES (?)`, [position]);
+
+    const rows = await sql.get(
+      `SELECT last_insert_rowid() AS id FROM category`,
+    );
+    const { id } = rows[0];
+
+    return NextResponse.json({ message: "Category created", id });
+  } catch (error) {
+    return NextResponse.json({ error }, { status: 401 });
+  }
+}
+
+export async function DELETE(request: NextRequest) {
+  try {
+    await Auth.check(request);
+
+    const url = request.nextUrl;
+    const id = url.searchParams.get("id");
+
+    if (!id) {
+      return NextResponse.json(
+        { error: "Category not found" },
+        { status: 404 },
+      );
+    }
+
+    await sql.execute(`DELETE FROM category WHERE id = ?`, [id]);
+
+    return NextResponse.json({ message: "Category deleted" });
+  } catch (error) {
+    return NextResponse.json({ error }, { status: 401 });
+  }
+}
+
+export async function PATCH(request: NextRequest) {
+  try {
+    await Auth.check(request);
+
+    const { id, name, color, position } = await request.json();
+
+    if (name !== undefined) {
+      await sql.execute(
+        `UPDATE category
+         SET name = ?
+         WHERE id = ?`,
+        [name, id],
+      );
+    }
+
+    if (color !== undefined) {
+      await sql.execute(
+        `UPDATE category
+         SET color = ?
+         WHERE id = ?`,
+        [color, id],
+      );
+    }
+
+    if (position !== undefined) {
+      await sql.execute(
+        `UPDATE category
+         SET position = ?
+         WHERE id = ?`,
+        [position, id],
+      );
+    }
+
+    return NextResponse.json({ message: "Category updated" });
+  } catch (error) {
+    return NextResponse.json({ error }, { status: 401 });
+  }
+}