+++ src/services/api.ts
@@ -1,5 +1,6 @@
 import * as Fetcher from "@/services/fetcher";
 import type { PostItItemPatch } from "@/@types/post-it";
+import type { CategoryItem } from "@/@types/category";
 
 const apiUrl = process.env.API_URL as string;
 
@@ -44,4 +45,23 @@ export async function createPostIt(
 
 export async function getCategories(authorization: string) {
   return Fetcher.GET(`${apiUrl}/api/categories`, undefined, { authorization });
+}
+
+export async function createCategory(authorization: string, position: number) {
+  return Fetcher.POST(`${apiUrl}/api/category`, { position }, undefined, {
+    authorization,
+  });
+}
+
+export async function updateCategory(
+  authorization: string,
+  category: Partial<CategoryItem>,
+) {
+  return Fetcher.PATCH(`${apiUrl}/api/category`, category, undefined, {
+    authorization,
+  });
+}
+
+export async function removeCategory(authorization: string, id: number) {
+  return Fetcher.DELETE(`${apiUrl}/api/category`, { id }, { authorization });
 }